Make warnings user-facing
