+++ crates/puffin-resolver/Cargo.toml
@@ -25,7 +25,7 @@ distribution-types = { path = "../distribution-types" }
 puffin-distribution = { path = "../puffin-distribution" }
 puffin-git = { path = "../puffin-git" }
 puffin-interpreter = { path = "../puffin-interpreter" }
-puffin-macros = { path = "../puffin-macros" }
+puffin-warnings = { path = "../puffin-warnings" }
 puffin-normalize = { path = "../puffin-normalize" }
 puffin-traits = { path = "../puffin-traits" }
 pypi-types = { path = "../pypi-types" }
