+++ crates/puffin-resolver/src/pubgrub/dependencies.rs
@@ -5,7 +5,6 @@ use tracing::warn;
 
 use pep508_rs::{MarkerEnvironment, Requirement, VersionOrUrl};
 use puffin_cache::CanonicalUrl;
-use puffin_macros::warn_once;
 use puffin_normalize::{ExtraName, PackageName};
 
 use crate::pubgrub::specifier::PubGrubSpecifier;
@@ -31,7 +30,7 @@ impl PubGrubDependencies {
         for requirement in requirements {
             // Avoid self-dependencies.
             if source.is_some_and(|source| source == &requirement.name) {
-                warn_once!("{} has a dependency on itself", requirement.name);
+                warn!("{} has a dependency on itself", requirement.name);
                 continue;
             }
 
@@ -78,7 +77,7 @@ impl PubGrubDependencies {
         for constraint in constraints {
             // Avoid self-dependencies.
             if source.is_some_and(|source| source == &constraint.name) {
-                warn_once!("{} has a dependency on itself", constraint.name);
+                warn!("{} has a dependency on itself", constraint.name);
                 continue;
             }
